+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## swamp v0.8.3
 
+* `-quiet` suppresses output
 * more explicit error messages for windows users
 
 ## swamp v0.8.2